requirements.txt
Django>=3.2,<4.0
djangorestframework
beautifulsoup4
celery
redis
kombu
requests
selenium
